# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tema": {
      "post": {
        "operationId": "TemaController_post",
        "summary": "Crear un nuevo tema",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El tema ha sido creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "tema"
        ]
      },
      "get": {
        "operationId": "TemaController_getAll",
        "summary": "Obtener todos los temas",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": " query-Filter.e.g.col1:v1,col2:v2… ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve todos los temas.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemaDTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
    "/tema/{id}": {
      "get": {
        "operationId": "TemaController_getById",
        "summary": "Obtener un tema por Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el tema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemaDTO"
                }
              }
            }
          },
          "404": {
            "description": "Tema no encontrado."
          }
        },
        "tags": [
          "tema"
        ]
      },
      "put": {
        "operationId": "TemaController_put",
        "summary": "Actualizar un tema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El tema ha sido actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemaDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Tema no encontrado."
          }
        },
        "tags": [
          "tema"
        ]
      },
      "delete": {
        "operationId": "TemaController_delete",
        "summary": "Eliminar un tema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El tema ha sido eliminado exitosamente."
          },
          "404": {
            "description": "Tema no encontrado."
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
    "/tema/{temaId}/subtemas": {
      "post": {
        "operationId": "TemaController_agregarSubtema",
        "summary": "Agregar un subtema a un tema",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtemaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El subtema ha sido agregado exitosamente."
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
    "/tema/{temaId}/subtemas/{subtemaId}": {
      "put": {
        "operationId": "TemaController_actualizarSubtema",
        "summary": "Actualizar un subtema",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtemaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El subtema ha sido actualizado exitosamente."
          }
        },
        "tags": [
          "tema"
        ]
      },
      "delete": {
        "operationId": "TemaController_eliminarSubtema",
        "summary": "Eliminar un subtema",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtemaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El subtema ha sido eliminado exitosamente."
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
    "/tema/{temaId}/subtemas/{subtemaId}/hallazgos": {
      "post": {
        "operationId": "TemaController_agregarHallazgo",
        "summary": "Agregar un hallazgo a un subtema",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtemaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HallazgoDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El hallazgo ha sido agregado exitosamente."
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
    "/tema/{temaId}/subtemas/{subtemaId}/hallazgos/{hallazgoId}": {
      "put": {
        "operationId": "TemaController_actualizarHallazgo",
        "summary": "Actualizar un hallazgo",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtemaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hallazgoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HallazgoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El hallazgo ha sido actualizado exitosamente."
          }
        },
        "tags": [
          "tema"
        ]
      },
      "delete": {
        "operationId": "TemaController_eliminarHallazgo",
        "summary": "Eliminar un hallazgo",
        "parameters": [
          {
            "name": "temaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtemaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hallazgoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El hallazgo ha sido eliminado exitosamente."
          }
        },
        "tags": [
          "tema"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan AnualAuditoria CRUD",
    "description": "API CRUD para la gestión de las auditoría y planes de mejoramiento",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TemaDTO": {
        "type": "object",
        "properties": {
          "informe_id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "subtema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fecha_creacion": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "informe_id",
          "titulo",
          "activo",
          "subtema",
          "fecha_creacion"
        ]
      },
      "SubtemaDTO": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "hallazgo": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "titulo",
          "activo",
          "hallazgo"
        ]
      },
      "HallazgoDTO": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string"
          },
          "criterio": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        },
        "required": [
          "titulo",
          "criterio",
          "descripcion",
          "activo"
        ]
      },
...
    }
  }
}